fix: split store into module
